+++ ui/src/components/Hoc/index.ts
@@ -1,5 +1,2 @@
 export { default as withSize } from "./withSize";
 export { default as withLabelSize } from "./withLabelSize";
-export { default as withPageTitle } from "./withPageTitle";
-
-export { default as usePageTitle } from "./withPageTitle/usePageTitle";